+++ supvisors/tests/test_supvisorsctl.py
@@ -814,9 +814,9 @@ def test_lazy_update_numprocs(controller, plugin, mocked_check):
     assert mocked_check.call_args_list == [call()]
     mocked_check.reset_mock()
     # test help and request
-    mocked_rpc = plugin.supvisors().lazy_update_numprocs
+    mocked_rpc = plugin.supvisors().update_numprocs
     _check_call(controller, mocked_check, mocked_rpc, plugin.help_lazy_update_numprocs, plugin.do_lazy_update_numprocs,
-                'dummy_process 2', [call('dummy_process', 2)])
+                'dummy_process 2', [call('dummy_process', 2, True, True)])
 
 
 def test_enable(controller, plugin, mocked_check):